Rename registry module to connection
